feat: Change router names and components to appropriate ones.
