+++ src/front-end/src/components/Login.js
@@ -1,40 +1,103 @@
 import React, {Component} from "react";
+import axios from "axios";
 import Webcam from "react-webcam";
 import "../App.css";
-import {Spinner, Button, Label} from "reactstrap";
-import {Link, BrowserRouter as Router, Route, Switch} from "react-router-dom";
+import {Spinner, Button} from "reactstrap";
+import {
+	withRouter,
+	Link,
+	BrowserRouter as Router,
+	Route,
+	Switch
+} from "react-router-dom";
+import "base64-to-image";
 
-class Login3 extends Component {
+class Login extends Component {
 	state = {
-		userName: ""
+		userFace: null
 	};
-	userRegister = e => {
-		this.setState({
-			userName: e.target.value
-		});
+
+	componentDidMount(Webcam) {
+		this.getLogin();
+	}
+
+	setRef = webcam => {
+		this.webcam = webcam;
 	};
+
+	faceDetected() {
+		this.props.history.push("/Option");
+		console.log("얼굴 정보 있음, 로그인 4 페이지로 넘어감");
+	}
+
+	faceNotDetected() {
+		this.props.history.push("/Signup", {userFace: this.state.userFace});
+		console.log("얼굴 정보 없음, 로그인 3 페이지로 넘어감");
+	}
+
+	getLogin = async () => {
+		console.log("캡처되고있음");
+		const dataURLtoFile = (dataurl, filename) => {
+			var arr = dataurl.split(","),
+				mime = arr[0].match(/:(.*?);/)[1],
+				bstr = atob(arr[1]),
+				n = bstr.length,
+				u8arr = new Uint8Array(n);
+
+			while (n--) {
+				u8arr[n] = bstr.charCodeAt(n);
+			}
+
+			return new File([u8arr], filename, {type: mime});
+		};
+
+		const captureImg = setTimeout(() => {
+			var base64Str = this.webcam.getScreenshot();
+			var file = dataURLtoFile(base64Str, "hello.jpg");
+			console.log(file);
+			console.log("캡처됨");
+			this.setState({
+				userFace: file
+			});
+			this.userFace();
+		}, 5000);
+	};
+
+	userFace = async () => {
+		let form_data = new FormData();
+		form_data.append("userFace", this.state.userFace);
+		try {
+			const response = await axios.post("api/v1/login/", form_data, {
+				headers: {
+					"content-type": "multipart/form-data"
+				}
+			});
+			console.log(response);
+			this.faceDetected();
+		} catch (error) {
+			console.error(error);
+			this.faceNotDetected();
+		}
+	};
+
 	render() {
 		return (
 			<div class="container-fluid">
 				<div class="row">
 					<div class="col-4" id="login">
-						<>
-							<Webcam
-								class="webcam"
-								audio={false}
-								facingmode="user"
-								ref={this.loginRef}
-								screenshotFormat="image/jpeg"
-							/>
+						<Webcam
+							class="webcam"
+							id="blink"
+							audio={false}
+							facingmode="user"
+							ref={this.setRef}
+							screenshotFormat="image/jpeg"
+						/>
 
-							<div
-								class="alert alert-secondary border-0 "
-								id="text"
-								role="alert"
-							>
-								<strong>[안내]</strong> 5초 후 화면이 캡처됩니다.
-							</div>
-						</>
+						<Spinner onLoad={this.capture} color="secondary" id="spinner" />
+						<div class="alert alert-secondary border-0 " id="text" role="alert">
+							<strong>[안내]</strong> 5초 후 화면이 캡처됩니다.
+						</div>
 					</div>
 
 					<div class="col-8" id="explain">
@@ -107,4 +170,4 @@ class Login3 extends Component {
 	}
 }
 
-export default Login3;
+export default withRouter(Login);
